# Conflicts:
#	CHANGELOG.md
#	cnccode/business_classes/BUSageExport.inc.php
#	cnccode/controller_classes/CTContact.inc.php
#	cnccode/controller_classes/CTCustomer.inc.php
#	cnccode/controller_classes/CTCustomerCRM.inc.php
#	cnccode/controller_classes/CTCustomerNote.inc.php
#	cnccode/controller_classes/CTHeader.inc.php
#	cnccode/controller_classes/CTSector.inc.php
#	cnccode/customClasses/QuestionnaireReportGenerator.php
#	cnccode/data_classes/DBEContact.inc.php
#	cnccode/data_classes/DBESite.inc.php
#	cnccode/templates/ContactEdit.inc.html
#	cnccode/templates/ContactReport.inc.html
#	cnccode/templates/CustomerCRM.inc.html
#	cnccode/templates/CustomerEditContact.html
#	cnccode/templates/CustomerEditSimple.inc.html
#	cnccode/templates/HeaderEdit.inc.html
#	composer.lock
#	dbVersioning/db.changelog-master.xml
#	htdocs/OutlookAddressEmail.php
#	htdocs/ProcessLeavers.php
#	htdocs/components/package.json
#	one-off-scripts/ContactImport.php

+++ htdocs/components/services/APIMain.js
@@ -1,7 +1,7 @@
 class APIMain {
     get(url) {
         return fetch(url)
-            .then((res) => res.json());
+        .then((res) => this.handleResponse(res));
     }
 
     getCustomers() {
@@ -14,14 +14,20 @@ class APIMain {
             .then((res) => res.json());
     }
 
-    uploadFiles(url, files, name) {
+    uploadFiles(url, files, name, data = null, handleException = false) {
         let payload = new FormData();
-        for (const file of files)
-            payload.append(name, file);
-        return fetch(url, {
-            method: "POST",
-            body: payload
-        });
+        for (const file of files) payload.append(name, file);
+        if (data) payload.append("data", JSON.stringify(data));
+        if (!handleException)
+            return fetch(url, {
+                method: "POST",
+                body: payload,
+            });
+        else
+            return fetch(url, {
+                method: "POST",
+                body: payload,
+            }).then((res) => this.handleResponse(res));
     }
 
     post(url, payload) {
@@ -31,17 +37,38 @@ class APIMain {
         });
     }
 
-    put(url, payload) {
+    postJson(url, payload) {
         return fetch(url, {
-            method: "PUT",
+            method: "POST",
             body: JSON.stringify(payload)
         }).then((res) => res.json());
     }
-    delete(url ) {
-        return fetch(url, {
-            method: "DELETE",            
-        }).then((res) => res.json());
+
+    put(url, payload, handleException = false) {
+
+        if (!handleException)
+            return fetch(url, {
+                method: "PUT",
+                body: JSON.stringify(payload),
+            }).then((res) => res.json());
+        else
+            return fetch(url, {
+                method: "PUT",
+                body: JSON.stringify(payload),
+            }).then((res) => this.handleResponse(res));
     }
+
+    delete(url,handleException=false) {
+      if (!handleException)
+        return fetch(url, {
+            method: "DELETE",
+        }).then((res) => res.json());
+        else
+        return fetch(url, {
+          method: "DELETE",
+      }).then((res) => this.handleResponse(res));
+    }
+
     postFormData(url, payload) {
         return fetch(url, {
             method: "POST",
@@ -49,6 +76,24 @@ class APIMain {
         });
     }
 
+    handleResponse = async (response) => {
+        const statusCode = response.status;
+        const text = await response.text();
+        return new Promise((res, reject) => {
+            try {
+                const textJson = JSON.parse(text);
+                textJson.status = statusCode;
+                if (response.ok) return res(textJson);
+                else {
+                    textJson.status = statusCode;
+                    return reject(textJson);
+                }
+            } catch (ex) {
+                return reject({state: false, error: "Data not saved successfully", responseCode: 400});
+            }
+        })
+
+    }
 }
 
 export default APIMain;
